src/requirements.txt
amqp==2.2.2
attrs==17.4.0
awesome-slugify==1.6.5
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
coverage==4.5.1
decorator==4.2.1
defusedxml==0.5.0
Django==2.0.2
django-allauth==0.35.0
django-behaviors==0.4.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-filter==1.1.0
django-rest-auth==0.9.3
django-suit==0.2.26
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
Faker==0.8.8
flake8==3.5.0
flake8-isort==2.4
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.11.1
kombu==4.1.0
mccabe==0.6.1
mixer==6.0.1
oauthlib==2.0.6
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
py==1.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.6.0
pytest==3.4.1
pytest-cov==2.5.1
pytest-django==3.1.2
python-dateutil==2.7.0
python3-openid==3.1.0
pytz==2018.3
regex==2018.2.21
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
testfixtures==5.4.0
text-unidecode==1.2
traitlets==4.3.2
Unidecode==0.4.21
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
